pyproject.toml
[project]
name = "openmetadata-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
